# contracts/openapi/latest/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "kind": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "LocalizedText": {
        "additionalProperties": {
          "type": "string"
        },
        "type": "object"
      },
...
      "UIActionDefinition": {
        "additionalProperties": true,
        "properties": {
          "custom_entry_key": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "label_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "render_mode": {
            "type": "string"
          },
          "required_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "route_path": {
            "type": "string"
          },
          "view_key": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "UIBootstrapResponse": {
        "example": {
          "available_surfaces": [
            "backoffice",
            "self_service"
          ],
          "default_path": "/self-service/requests",
          "self_service_apis": [
            "documents.self_service.requests.create",
            "documents.self_service.requests.get",
            "documents.self_service.requests.list",
            "documents.self_service.requests.submit",
            "leave.self_service.balances.list",
            "leave.self_service.requests.amend",
            "leave.self_service.requests.cancel",
            "leave.self_service.requests.create",
            "leave.self_service.requests.get",
            "leave.self_service.requests.list",
            "leave.self_service.requests.submit",
            "leave.self_service.requests.update"
          ],
          "surface": "self_service"
        },
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/UIActionDefinition"
            },
            "type": "array"
          },
          "admin_access": {
            "type": "boolean"
          },
          "admin_path": {
            "type": "string"
          },
          "available_surfaces": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "default_path": {
            "type": "string"
          },
          "flows": {
            "items": {
              "additionalProperties": true,
              "description": "UI document flow",
              "type": "object"
            },
            "type": "array"
          },
          "locale": {
            "type": "string"
          },
          "menus": {
            "items": {
              "$ref": "#/components/schemas/UIMenuDefinition"
            },
            "type": "array"
          },
          "self_service_apis": {
            "items": {
              "$ref": "#/components/schemas/UISelfServiceAPIDefinition"
            },
            "type": "array"
          },
          "supported_locales": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "surface": {
            "type": "string"
          },
          "views": {
            "items": {
              "$ref": "#/components/schemas/UIViewDefinition"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "UICardDefinition": {
        "additionalProperties": true,
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "label_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UIMenuDefinition": {
        "properties": {
          "action_key": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "label_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "order": {
            "type": "integer"
          },
          "required_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "UISectionDefinition": {
        "additionalProperties": true,
        "properties": {
          "fields": {
            "items": {
              "additionalProperties": true,
              "description": "Section field",
              "type": "object"
            },
            "type": "array"
          },
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object"
      },
      "UISelfServiceAPIDefinition": {
        "additionalProperties": true,
        "properties": {
          "audience_kinds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "description_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "document_type": {
            "type": "string"
          },
          "flow_key": {
            "type": "string"
          },
          "handler_kind": {
            "type": "string"
          },
          "idempotent": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "model_key": {
            "type": "string"
          },
          "offline_capable": {
            "type": "boolean"
          },
          "request_contract_key": {
            "type": "string"
          },
          "required_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "response_contract_key": {
            "type": "string"
          },
          "route_path": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object"
      },
      "UISelfServiceAPIListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UISelfServiceAPIDefinition"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "UITabDefinition": {
        "properties": {
          "key": {
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/UISectionDefinition"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "title_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object"
      },
      "UIViewColumn": {
        "additionalProperties": true,
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "label_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "path": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "UIViewDefinition": {
        "additionalProperties": true,
        "properties": {
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/UICardDefinition"
            },
            "type": "array"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/UIViewColumn"
            },
            "type": "array"
          },
          "document_type": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "model_key": {
            "type": "string"
          },
          "projection_key": {
            "type": "string"
          },
          "required_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tabs": {
            "items": {
              "$ref": "#/components/schemas/UITabDefinition"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "title_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI document for the runtime generic platform APIs. Registered modules become available through the generic model, document, UI contract, search, and offline endpoints. Custom module-specific controller namespaces are not auto-generated.",
    "title": "Orbyte Platform API",
    "version": "0.1.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ui/bootstrap": {
      "get": {
        "description": "Returns the surface-aware workspace menus, actions, views, locale, default route path, and self-service API catalog for the current principal.",
        "parameters": [
          {
            "description": "Optional workspace surface key.",
            "in": "query",
            "name": "surface",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIBootstrapResponse"
                }
              }
            },
            "description": "UI bootstrap"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Get UI bootstrap",
        "tags": [
          "UI Contracts"
        ],
        "x-orbyte-support-level": "public-headless"
      }
    },
...
    "/ui/self-service/apis": {
      "get": {
        "description": "Returns the explicit self-service APIs the current principal is allowed to invoke.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UISelfServiceAPIListResponse"
                }
              }
            },
            "description": "Self-service API list"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "List self-service APIs",
        "tags": [
          "UI Contracts"
        ],
        "x-orbyte-support-level": "public-headless"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}